feat: Cleanup dockerfiles
