+++ apps/BingBot/requirements.txt
@@ -1,4 +1,4 @@
 discord.py==2.3.2
-python-dotenv==0.21.1
-PyYAML==6.0
-EdgeGPT==0.13.2
+python-dotenv==1.0.0
+PyYAML==6.0.1
+bing-chat==1.9.3
